requirements.txt
flask
pandas
numpy
scipy
sklearn
Pillow
tensorflow
gunicorn
